+++ src/commands/env/create/compute.ts
@@ -10,7 +10,6 @@ import { Flags } from '@oclif/core';
 import { Messages } from '@salesforce/core';
 import { QueryResult } from 'jsforce';
 import { cli } from 'cli-ux';
-import { format } from 'date-fns';
 import Command from '../../../lib/base';
 import { FunctionsFlagBuilder } from '../../../lib/flags';
 import pollForResult from '../../../lib/poll-for-result';
@@ -37,28 +36,19 @@ export default class EnvCreateCompute extends Command {
     alias: Flags.string({
       char: 'a',
       description: messages.getMessage('flags.alias.summary'),
-      exclusive: ['setalias'],
     }),
-
-    setalias: Flags.string({
-      char: 'a',
-      description: messages.getMessage('flags.alias.summary'),
-      exclusive: ['alias'],
-      hidden: true,
-    }),
+    json: FunctionsFlagBuilder.json,
   };
 
   async run() {
     const { flags } = await this.parse(EnvCreateCompute);
+    this.postParseHook(flags);
 
-    const alias = flags.alias ?? flags.setalias;
-
-    if (flags.setalias) {
-      this.warn(messages.getMessage('flags.setalias.deprecation'));
-    }
+    const alias = flags.alias;
 
     // if `--connected-org` is null here, fetchOrg will pull the default org from the surrounding environment
     const org = await fetchOrg(flags['connected-org']);
+
     const orgId = org.getOrgId();
 
     if (!(await this.isFunctionsEnabled(org))) {
@@ -74,9 +64,9 @@ export default class EnvCreateCompute extends Command {
           )} to the "features" list in your scratch org definition JSON file, e.g. "features": ["Functions"]`
       );
     }
-
-    cli.action.start(`Creating compute environment for org ID ${orgId}`);
-
+    if (!flags.json) {
+      cli.action.start(`Creating compute environment for org ID ${orgId}`);
+    }
     const project = await fetchSfdxProject();
     const projectName = project.name;
 
@@ -85,7 +75,6 @@ export default class EnvCreateCompute extends Command {
     }
 
     const connection = org.getConnection();
-
     await pollForResult(async () => {
       // This query allows us to verify that the org connection has actually been created before
       // attempting to create a compute environment. If we don't wait for this to happen, environment
@@ -97,13 +86,13 @@ export default class EnvCreateCompute extends Command {
           Id,
           Status,
           Error
-          FROM SfFunctionsConnection`);
+          FROM FunctionsConnection`);
       } catch (err) {
         const error = err as Error;
-        // This is obviously heinous, but should only exist short-term until the move from `SfFunctionsConnection`
+        // This is obviously heinous, but should only exist short-term until the move from `FunctionsConnection`
         // to `FunctionConnection` is complete. Once that's done, we can remove this and go back to a simple
         // query against `FunctionConnection`
-        if (!error.message.includes("sObject type 'SfFunctionsConnection' is not supported.")) {
+        if (!error.message.includes("sObject type 'FunctionsConnection' is not supported.")) {
           this.error(error);
         }
         response = await connection.query<FunctionConnectionRecord>(`SELECT
@@ -131,7 +120,7 @@ export default class EnvCreateCompute extends Command {
       // If there is any other error besides the one mentioned above, something is actually wrong
       // and we should bail
       if (record.Error) {
-        this.error(record.Error);
+        this.error(`${record.Error}`);
       }
 
       // This is the go signal. Once we have this status it means that the connection is fully up
@@ -151,42 +140,59 @@ export default class EnvCreateCompute extends Command {
 
       cli.action.stop();
 
-      this.log(`New compute environment created with ID ${app.name}`);
+      if (!flags.json) {
+        this.log(`New compute environment created with ID ${app.name}`);
 
-      cli.action.start('Connecting environments');
+        cli.action.start('Connecting environments');
+      }
 
       if (alias) {
-        this.info.aliases.set(alias, app.id!);
-
-        await this.info.write();
+        this.stateAggregator.aliases.set(alias, app.id!);
+        await this.stateAggregator.aliases.write();
       }
 
       cli.action.stop();
-
-      this.log(
-        alias
-          ? `Your compute environment with local alias ${herokuColor.cyan(alias)} is ready.`
-          : 'Your compute environment is ready.'
-      );
+      if (!flags.json) {
+        this.log(
+          alias
+            ? `Your compute environment with local alias ${herokuColor.cyan(alias)} is ready.`
+            : 'Your compute environment is ready.'
+        );
+      }
     } catch (err) {
-      const DUPLICATE_PROJECT_MESSAGE = 'This org is already connected to a compute environment for this project';
+      const DUPLICATE_PROJECT_MESSAGE =
+        'There is already a project with the same name in the same namespace for this org';
+      const INVALID_PROJECT_NAME =
+        "Sfdx project name may only contain numbers (0-9), letters (a-z A-Z) and non-consecutive underscores ('_'). It must begin with a letter and end with either a number or letter.";
       const error = err as { data: { message?: string } };
+      cli.action.stop('error!');
+
+      if (error.data?.message?.includes(INVALID_PROJECT_NAME)) {
+        this.error(
+          "Project name may only contain numbers (0-9), letters (a-z A-Z) and non-consecutive underscores ('_'). It must begin with a letter and end with either a number or letter"
+        );
+      }
       // If environment creation fails because an environment already exists for this org and project
       // we want to fetch the existing environment so that we can point the user to it
       if (error.data?.message?.includes(DUPLICATE_PROJECT_MESSAGE)) {
-        cli.action.stop('error!');
         const app = await fetchAppForProject(this.client, projectName, org.getUsername());
-
-        this.log(`${DUPLICATE_PROJECT_MESSAGE}:`);
-        this.log(`Compute Environment ID: ${app.name}`);
-        if (app.created_at) {
-          this.log(`Created on: ${format(new Date(app.created_at), 'E LLL d HH:mm:ss O y')}`);
-        }
-
-        return;
+        this.error(`This org is already connected to a compute environment for this project -> ${app.name}`);
       }
-
-      throw error;
+      this.error(`${error.data.message}`);
+    }
+    const app = await fetchAppForProject(this.client, projectName, org.getUsername());
+    if (flags.json) {
+      cli.styledJSON({
+        status: 0,
+        result: {
+          alias,
+          projectName,
+          connectedOrgAlias: '',
+          connectedOrgId: orgId,
+          computeEnvironmentName: app.name,
+        },
+        warnings: [],
+      });
     }
   }
 }