+++ src/commands/env/logdrain/add.ts
@@ -43,10 +43,13 @@ export default class LogDrainAdd extends Command {
       description: messages.getMessage('flags.drain-url.summary'),
       hidden: true,
     }),
+    json: FunctionsFlagBuilder.json,
   };
 
   async run() {
     const { flags } = await this.parse(LogDrainAdd);
+    this.postParseHook(flags);
+
     // We support both versions of the flag here for the sake of backward compat
     const targetCompute = flags['target-compute'] ?? flags.environment;
     const url = flags['drain-url'] ?? flags.url;
@@ -54,7 +57,7 @@ export default class LogDrainAdd extends Command {
     if (!targetCompute) {
       throw new Errors.CLIError(
         `Missing required flag:
-        -c, --target-compute TARGET-COMPUTE  ${herokuColor.dim('Environment name.')}
+        -e, --target-compute TARGET-COMPUTE  ${herokuColor.dim('Environment name.')}
        See more help with --help`
       );
     }
@@ -68,23 +71,65 @@ export default class LogDrainAdd extends Command {
     }
 
     if (flags.environment) {
-      this.warn(messages.getMessage('flags.environment.deprecation'));
+      cli.warn(messages.getMessage('flags.environment.deprecation'));
     }
 
     if (flags.url) {
-      this.warn(messages.getMessage('flags.url.deprecation'));
+      cli.warn(messages.getMessage('flags.url.deprecation'));
     }
-
     const appName = await resolveAppNameForEnvironment(targetCompute);
 
-    cli.action.start(`Creating drain for environment ${herokuColor.app(targetCompute)}`);
+    try {
+      const result = await this.client.post<Heroku.LogDrain>(`/apps/${appName}/log-drains`, {
+        data: {
+          url,
+        },
+      });
 
-    await this.client.post<Heroku.LogDrain>(`/apps/${appName}/log-drains`, {
-      data: {
-        url,
-      },
-    });
+      if (flags.json) {
+        cli.styledJSON({
+          status: 0,
+          result: [
+            {
+              addon: null,
+              created_at: result.data.created_at,
+              id: result.data.id,
+              token: result.data.token,
+              updated_at: result.data.updated_at,
+              url: result.data.url,
+            },
+          ],
+          warnings: [],
+        });
+      } else {
+        cli.action.start(`Creating drain for environment ${herokuColor.app(targetCompute)}`);
 
-    cli.action.stop();
+        cli.action.stop();
+      }
+    } catch (e) {
+      const error = e as { data: { message?: string } };
+
+      if (error.data?.message?.includes('Url is invalid')) {
+        this.error(new Error(`URL is invalid <${url}>`));
+      }
+
+      if (error.data?.message?.includes('Url has already been taken')) {
+        this.error(new Error(`Logdrain URL is already added <${url}>`));
+      }
+
+      if (error.data?.message?.includes("Couldn't find that app.")) {
+        this.error(new Error(`Could not find environment <${appName}>`));
+      }
+
+      if (error.data?.message?.includes("You've reached the limit")) {
+        this.error(new Error(`You've reached the limit of 5 log drains on <${appName}>`));
+      }
+
+      if (error.data?.message?.includes('401')) {
+        this.error(new Error('Your token has expired, please login with sf login functions'));
+      }
+
+      this.error(e as Error);
+    }
   }
 }