+++ src/commands/env/var/get.ts
@@ -8,6 +8,7 @@ import herokuColor from '@heroku-cli/color';
 import * as Heroku from '@heroku-cli/schema';
 import { Messages } from '@salesforce/core';
 import { Errors } from '@oclif/core';
+import { cli } from 'cli-ux';
 import { FunctionsFlagBuilder } from '../../../lib/flags';
 
 import Command from '../../../lib/base';
@@ -32,6 +33,7 @@ export default class VarGet extends Command {
       exclusive: ['target-compute'],
       hidden: true,
     }),
+    json: FunctionsFlagBuilder.json,
   };
 
   static args = [
@@ -43,19 +45,21 @@ export default class VarGet extends Command {
 
   async run() {
     const { flags, args } = await this.parse(VarGet);
+    this.postParseHook(flags);
+
     // We support both versions of the flag here for the sake of backward compat
     const targetCompute = flags['target-compute'] ?? flags.environment;
 
     if (!targetCompute) {
       throw new Errors.CLIError(
         `Missing required flag:
-        -c, --target-compute TARGET-COMPUTE  ${herokuColor.dim('Environment name.')}
+        -e, --target-compute TARGET-COMPUTE  ${herokuColor.dim('Environment name.')}
        See more help with --help`
       );
     }
 
     if (flags.environment) {
-      this.warn(messages.getMessage('flags.environment.deprecation'));
+      cli.warn(messages.getMessage('flags.environment.deprecation'));
     }
 
     const appName = await resolveAppNameForEnvironment(targetCompute);
@@ -64,14 +68,30 @@ export default class VarGet extends Command {
 
     const value = config[args.key];
 
-    if (!value) {
-      this.warn(
-        `No config var named ${herokuColor.cyan(args.key as string)} found for environment ${herokuColor.cyan(
-          targetCompute
-        )}`
-      );
-    }
+    if (flags.json) {
+      if (!value) {
+        cli.styledJSON({
+          status: 0,
+          result: null,
+          warnings: [`No config var named ${args.key as string} found for environment <${targetCompute}>`],
+        });
+        return;
+      }
 
-    this.log(value);
+      cli.styledJSON({
+        status: 0,
+        result: value,
+        warnings: [],
+      });
+    } else {
+      if (!value) {
+        cli.warn(
+          `No config var named ${herokuColor.cyan(args.key as string)} found for environment ${herokuColor.cyan(
+            targetCompute
+          )}`
+        );
+      }
+      this.log(value);
+    }
   }
 }