+++ src/commands/env/var/list.ts
@@ -38,19 +38,21 @@ export default class ConfigList extends Command {
 
   async run() {
     const { flags } = await this.parse(ConfigList);
+    this.postParseHook(flags);
+
     // We support both versions of the flag here for the sake of backward compat
     const targetCompute = flags['target-compute'] ?? flags.environment;
 
     if (!targetCompute) {
       throw new Errors.CLIError(
         `Missing required flag:
-        -c, --target-compute TARGET-COMPUTE  ${herokuColor.dim('Environment name.')}
+        -e, --target-compute TARGET-COMPUTE  ${herokuColor.dim('Environment name.')}
        See more help with --help`
       );
     }
 
     if (flags.environment) {
-      this.warn(messages.getMessage('flags.environment.deprecation'));
+      cli.warn(messages.getMessage('flags.environment.deprecation'));
     }
 
     const appName = await resolveAppNameForEnvironment(targetCompute);
@@ -81,7 +83,7 @@ export default class ConfigList extends Command {
       });
     } else {
       if (!configArray.length) {
-        this.warn(`No config vars found for environment ${targetCompute}`);
+        cli.warn(`No config vars found for environment ${targetCompute}`);
         return;
       }