+++ src/commands/logout/functions.ts
@@ -8,6 +8,7 @@ import { cli } from 'cli-ux';
 import { Messages } from '@salesforce/core';
 
 import Command from '../../lib/base';
+import { FunctionsFlagBuilder } from '../../lib/flags';
 
 Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('@salesforce/plugin-functions', 'logout.functions');
@@ -17,12 +18,26 @@ export default class Login extends Command {
 
   static examples = messages.getMessages('examples');
 
+  static flags = {
+    json: FunctionsFlagBuilder.json,
+  };
+
   async run() {
+    const { flags } = await this.parse(Login);
+    this.postParseHook(flags);
+
     cli.action.start(messages.getMessage('action.start'));
 
-    this.info.tokens.unset(Command.TOKEN_BEARER_KEY);
-    await this.info.write();
+    this.stateAggregator.tokens.unset(Command.TOKEN_BEARER_KEY);
+    await this.stateAggregator.tokens.write();
 
     cli.action.stop();
+    if (flags.json) {
+      cli.styledJSON({
+        status: 0,
+        result: [],
+        warnings: [],
+      });
+    }
   }
 }