+++ src/lib/base.ts
@@ -5,28 +5,31 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { URL } from 'url';
-import { Command as Base } from '@oclif/core';
-import { GlobalInfo, Org } from '@salesforce/core';
+import { Interfaces } from '@oclif/core';
+import { SfCommand } from '@salesforce/sf-plugins-core';
+import { StateAggregator, Org } from '@salesforce/core';
 import { cli } from 'cli-ux';
 import APIClient, { herokuClientApiUrl } from './api-client';
 import herokuVariant from './heroku-variant';
 import { SfdcAccount } from './sfdc-types';
 
-export default abstract class Command extends Base {
+export default abstract class Command extends SfCommand<any> {
   protected static TOKEN_BEARER_KEY = 'functions-bearer';
   protected static TOKEN_REFRESH_KEY = 'functions-refresh';
   // We want to implement `--json` on a per-command basis, so we disable the global json flag here
-  static disableJsonFlag = true;
+  static enableJsonFlag = false;
 
-  protected info!: GlobalInfo;
+  protected stateAggregator!: StateAggregator;
 
   private _client?: APIClient;
 
   private _auth?: string;
 
+  protected outputJSON = false;
+
   protected async init(): Promise<void> {
     await super.init();
-    this.info = await GlobalInfo.getInstance();
+    this.stateAggregator = await StateAggregator.getInstance();
   }
 
   protected get identityUrl(): URL {
@@ -37,7 +40,7 @@ export default abstract class Command extends Base {
   }
 
   protected get username() {
-    return this.info.tokens.get(Command.TOKEN_BEARER_KEY)?.user;
+    return this.stateAggregator.tokens.get(Command.TOKEN_BEARER_KEY)?.user;
   }
 
   protected resetClientAuth() {
@@ -52,7 +55,7 @@ export default abstract class Command extends Base {
       if (apiKey) {
         this._auth = apiKey;
       } else {
-        const token = this.info.tokens.get(Command.TOKEN_BEARER_KEY, true)?.token;
+        const token = this.stateAggregator.tokens.get(Command.TOKEN_BEARER_KEY, true)?.token;
 
         if (!token) {
           throw new Error(`Not authenticated. Please login with \`${this.config.bin} login functions\`.`);
@@ -138,7 +141,7 @@ export default abstract class Command extends Base {
     const confirmedValue = this.fetchConfirmationValue(name, confirm);
     if (name !== confirmedValue) {
       warningMessage = warningMessage || `This will delete the ${type} ${name}`;
-      this.warn(`${warningMessage}\nTo proceed, enter the ${type} name (${name}) again in the prompt below:`);
+      cli.warn(`${warningMessage}\nTo proceed, enter the ${type} name (${name}) again in the prompt below:`);
       // This is a workaround for cli-ux
       // & fancy-test stubbing issues
       // cli-ux mocks itself incorrectly (tbd why)
@@ -150,5 +153,31 @@ export default abstract class Command extends Base {
         this.error('Confirmation name does not match');
       }
     }
+  }
+
+  // This interface is copied from oclif/core
+  error(input: string | Error, options: { code?: string; exit: false } & Interfaces.PrettyPrintableError): void;
+  error(input: string | Error, options?: { code?: string; exit?: number } & Interfaces.PrettyPrintableError): never;
+  error(
+    input: string | Error,
+    options: { code?: string; exit?: number | false } & Interfaces.PrettyPrintableError = {}
+  ): void {
+    if (this.outputJSON) {
+      if (typeof input === 'string') input = new Error(input);
+      const { message, name } = input;
+      cli.styledJSON({
+        status: 1,
+        message,
+        name,
+        warnings: [],
+      });
+      cli.exit(1);
+    } else {
+      super.error(input, options as any);
+    }
+  }
+
+  protected postParseHook(flags: { [name: string]: string | boolean | number | string[] | undefined }) {
+    this.outputJSON = flags.json as boolean;
   }
 }