+++ test/commands/env/create/compute.test.ts
@@ -5,11 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { expect, test } from '@oclif/test';
-import { Org, SfdxProject } from '@salesforce/core';
-import { AliasAccessor } from '@salesforce/core/lib/globalInfo';
+import { Org, SfProject } from '@salesforce/core';
+import { AliasAccessor } from '@salesforce/core/lib/stateAggregator';
 import * as sinon from 'sinon';
 import { AuthStubs } from '../../../helpers/auth';
-import vacuum from '../../../helpers/vacuum';
 
 const APP_MOCK = {
   id: '1',
@@ -63,6 +62,18 @@ const PROJECT_MOCK_NO_NAME = {
 const ORG_ALIAS = 'my-cool-alias';
 const ENVIRONMENT_ALIAS = 'my-cool-environment';
 
+const jsonSuccess = {
+  status: 0,
+  result: {
+    alias: ENVIRONMENT_ALIAS,
+    projectName: PROJECT_CONFIG_MOCK.name,
+    connectedOrgAlias: '',
+    connectedOrgId: ORG_MOCK.id,
+    computeEnvironmentName: APP_MOCK.name,
+  },
+  warnings: [],
+};
+
 describe('sf env create compute', () => {
   const sandbox = sinon.createSandbox();
 
@@ -72,7 +83,7 @@ describe('sf env create compute', () => {
     .retries(3)
     .do(() => {
       sandbox.stub(Org, 'create' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
     })
     .finally(() => {
       sandbox.restore();
@@ -82,6 +93,8 @@ describe('sf env create compute', () => {
         .post(`/sales-org-connections/${ORG_MOCK.id}/apps`, {
           sfdx_project_name: PROJECT_CONFIG_MOCK.name,
         })
+        .reply(200, APP_MOCK)
+        .get(`/sales-org-connections/${ORG_MOCK.id}/apps/${PROJECT_CONFIG_MOCK.name}`)
         .reply(200, APP_MOCK);
     })
     .command(['env:create:compute'])
@@ -95,6 +108,7 @@ describe('sf env create compute', () => {
 
   const aliasSetSpy = sandbox.spy();
   const aliasWriteSpy = sandbox.spy();
+  // const setAliasSpy = sandbox.spy();
 
   test
     .stdout()
@@ -102,9 +116,9 @@ describe('sf env create compute', () => {
     .retries(3)
     .do(() => {
       orgStub = sandbox.stub(Org, 'create' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       sandbox.stub(AliasAccessor.prototype, 'set' as any).callsFake(aliasSetSpy);
-      AuthStubs.write.callsFake(aliasWriteSpy);
+      AuthStubs.aliasesWrite.callsFake(aliasWriteSpy);
     })
     .finally(() => {
       sandbox.restore();
@@ -114,10 +128,12 @@ describe('sf env create compute', () => {
         .post(`/sales-org-connections/${ORG_MOCK.id}/apps`, {
           sfdx_project_name: PROJECT_CONFIG_MOCK.name,
         })
+        .reply(200, APP_MOCK)
+        .get(`/sales-org-connections/${ORG_MOCK.id}/apps/${PROJECT_CONFIG_MOCK.name}`)
         .reply(200, APP_MOCK);
     })
     .command(['env:create:compute', '-o', `${ORG_ALIAS}`, '-a', `${ENVIRONMENT_ALIAS}`])
-    .it('creates a compute environment using and sets an alias using values passed in flags', (ctx) => {
+    .it('creates a compute environment and sets an alias using values passed in flags', (ctx) => {
       expect(ctx.stderr).to.contain(`Creating compute environment for org ID ${ORG_MOCK.id}`);
       expect(ctx.stdout).to.contain(`New compute environment created with ID ${APP_MOCK.name}`);
       expect(ctx.stdout).to.contain(`Your compute environment with local alias ${ENVIRONMENT_ALIAS} is ready`);
@@ -127,13 +143,14 @@ describe('sf env create compute', () => {
     });
 
   test
+    .stdout()
     .stderr()
     .retries(3)
     .do(() => {
       orgStub = sandbox.stub(Org, 'create' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       sandbox.stub(AliasAccessor.prototype, 'set' as any).callsFake(aliasSetSpy);
-      AuthStubs.write.callsFake(aliasWriteSpy);
+      AuthStubs.aliasesWrite.callsFake(aliasWriteSpy);
     })
     .finally(() => {
       sandbox.restore();
@@ -143,13 +160,16 @@ describe('sf env create compute', () => {
         .post(`/sales-org-connections/${ORG_MOCK.id}/apps`, {
           sfdx_project_name: PROJECT_CONFIG_MOCK.name,
         })
+        .reply(200, APP_MOCK)
+        .get(`/sales-org-connections/${ORG_MOCK.id}/apps/${PROJECT_CONFIG_MOCK.name}`)
         .reply(200, APP_MOCK);
     })
-    .command(['env:create:compute', '-o', `${ORG_ALIAS}`, '--setalias', `${ENVIRONMENT_ALIAS}`])
-    .it('will use setalias if passed using the old flag (not --alias)', (ctx) => {
-      expect(vacuum(ctx.stderr).replace(/\n[›»]/gm, '')).to.contain(
-        vacuum('--setalias is deprecated and will be removed in a future release. Please use --alias going forward.')
-      );
+    .command(['env:create:compute', '-o', `${ORG_ALIAS}`, '-a', `${ENVIRONMENT_ALIAS}`, '-j'])
+    .it('will show json output with success', (ctx) => {
+      const succJSON = JSON.parse(ctx.stdout);
+
+      expect(succJSON.status).to.deep.equal(jsonSuccess.status);
+      expect(succJSON.result).to.eql(jsonSuccess.result);
     });
 
   test
@@ -157,7 +177,7 @@ describe('sf env create compute', () => {
     .stderr()
     .do(() => {
       orgStub = sandbox.stub(Org, 'create' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK_NO_NAME);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK_NO_NAME);
     })
     .finally(() => {
       sandbox.restore();
@@ -173,7 +193,7 @@ describe('sf env create compute', () => {
     .stderr()
     .do(() => {
       orgStub = sandbox.stub(Org, 'create' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
     })
     .finally(() => {
       sandbox.restore();
@@ -184,31 +204,59 @@ describe('sf env create compute', () => {
           sfdx_project_name: PROJECT_CONFIG_MOCK.name,
         })
         .reply(422, {
-          message: 'This org is already connected to a compute environment for this project',
+          message: 'Sfdx project name There is already a project with the same name in the same namespace for this org',
         })
         .get(`/sales-org-connections/${ORG_MOCK.id}/apps/${PROJECT_CONFIG_MOCK.name}`)
         .reply(200, APP_MOCK);
     })
     .command(['env:create:compute'])
-    .it('displays an informative error message when environment already exists for a given project', (ctx) => {
-      expect(ctx.stderr).to.contain('error!');
-      expect(ctx.stdout).to.contain(`Compute Environment ID: ${APP_MOCK.name}`);
-    });
+    .catch((error) => {
+      expect(error.message).to.contain('This org is already connected to a compute environment for this project');
+    })
+    .it('displays an informative error message when environment already exists for a given project');
+
+  test
+    .stdout()
+    .stderr()
+    .do(() => {
+      orgStub = sandbox.stub(Org, 'create' as any).returns(ORG_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
+    })
+    .finally(() => {
+      sandbox.restore();
+    })
+    .nock('https://api.heroku.com', (api) => {
+      api
+        .post(`/sales-org-connections/${ORG_MOCK.id}/apps`, {
+          sfdx_project_name: PROJECT_CONFIG_MOCK.name,
+        })
+        .reply(422, {
+          message:
+            "Sfdx project name may only contain numbers (0-9), letters (a-z A-Z) and non-consecutive underscores ('_'). It must begin with a letter and end with either a number or letter.",
+        });
+    })
+    .command(['env:create:compute'])
+    .catch((error) => {
+      expect(error.message).to.contain(
+        "Project name may only contain numbers (0-9), letters (a-z A-Z) and non-consecutive underscores ('_'). It must begin with a letter and end with either a number or letter"
+      );
+    })
+    .it('displays an informative error message when the project name is invalid');
 
   test
     .stdout()
     .stderr()
     .retries(3)
     .do(() => {
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       sandbox.stub(AliasAccessor.prototype, 'set' as any).callsFake(aliasSetSpy);
-      AuthStubs.write.callsFake(aliasWriteSpy);
+      AuthStubs.aliasesWrite.callsFake(aliasWriteSpy);
     })
     .add('queryStub', () => {
       const queryStub = sinon.stub();
       queryStub
-        .withArgs(sinon.match('SfFunctionsConnection'))
-        .throws({ message: "sObject type 'SfFunctionsConnection' is not supported." });
+        .withArgs(sinon.match('FunctionsConnection'))
+        .throws({ message: "sObject type 'FunctionsConnection' is not supported." });
 
       queryStub.withArgs(sinon.match('FunctionConnection')).returns({
         records: [
@@ -240,10 +288,12 @@ describe('sf env create compute', () => {
         .post(`/sales-org-connections/${ORG_MOCK.id}/apps`, {
           sfdx_project_name: PROJECT_CONFIG_MOCK.name,
         })
+        .reply(200, APP_MOCK)
+        .get(`/sales-org-connections/${ORG_MOCK.id}/apps/${PROJECT_CONFIG_MOCK.name}`)
         .reply(200, APP_MOCK);
     })
     .command(['env:create:compute', '-o', `${ORG_ALIAS}`, '-a', `${ENVIRONMENT_ALIAS}`])
-    .it('still creates a compute env even if SfFunctionsConnection is not supported', (ctx) => {
+    .it('still creates a compute env even if FunctionsConnection is not supported', (ctx) => {
       expect(ctx.stderr).to.contain(`Creating compute environment for org ID ${ORG_MOCK.id}`);
       expect(ctx.stdout).to.contain(`New compute environment created with ID ${APP_MOCK.name}`);
       expect(ctx.stdout).to.contain(`Your compute environment with local alias ${ENVIRONMENT_ALIAS} is ready`);
@@ -260,7 +310,7 @@ describe('sf env create compute', () => {
     .stderr()
     .do(() => {
       orgStub = sandbox.stub(Org, 'create' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       CONN_MOCK.metadata.list = sinon.stub().throws('INVALID_TYPE');
     })
     .finally(() => {