+++ test/commands/env/delete.test.ts
@@ -5,8 +5,8 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { expect, test } from '@oclif/test';
-import { Org, SfdxProject } from '@salesforce/core';
-import { AliasAccessor } from '@salesforce/core/lib/globalInfo';
+import { Org, SfProject } from '@salesforce/core';
+import { AliasAccessor } from '@salesforce/core/lib/stateAggregator';
 import * as sinon from 'sinon';
 import * as Utils from '../../../src/lib/utils';
 import vacuum from '../../helpers/vacuum';
@@ -46,7 +46,7 @@ describe('env:delete', () => {
     .nock('https://api.heroku.com', (api) => api.get(`/apps/${COMPUTE_ENV_NAME}`).reply(200))
     .do(() => {
       sandbox.stub(Utils, 'resolveOrg' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       sandbox.stub(Utils, 'findOrgExpirationStatus' as any).returns(false);
     })
     .finally(() => {
@@ -60,9 +60,30 @@ describe('env:delete', () => {
 
   test
     .stderr()
+    .stdout()
+    .nock('https://api.heroku.com', (api) => api.delete(`/apps/${COMPUTE_ENV_NAME}`).reply(200))
+    .nock('https://api.heroku.com', (api) => api.get(`/apps/${COMPUTE_ENV_NAME}`).reply(200))
     .do(() => {
       sandbox.stub(Utils, 'resolveOrg' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(Utils, 'findOrgExpirationStatus' as any).returns(false);
+    })
+    .finally(() => {
+      sandbox.restore();
+    })
+    .command(['env:delete', `--target-compute=${COMPUTE_ENV_NAME}`, `--confirm=${COMPUTE_ENV_NAME}`, '--json'])
+    .it('deletes an environment and returns json', (ctx) => {
+      expect(ctx.stderr).to.include(`Deleting environment ${COMPUTE_ENV_NAME}... done`);
+      expect(vacuum(ctx.stdout).replace(/\n[›»]/gm, '')).to.contain(
+        vacuum('{\n"status": 0,\n"result": "Environment deleted.",\n"warnings": []\n}')
+      );
+    });
+
+  test
+    .stderr()
+    .do(() => {
+      sandbox.stub(Utils, 'resolveOrg' as any).returns(ORG_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       sandbox.stub(AliasAccessor.prototype, 'getValue').returns(COMPUTE_ENV_NAME);
       sandbox.stub(Utils, 'findOrgExpirationStatus' as any).returns(false);
     })
@@ -82,10 +103,10 @@ describe('env:delete', () => {
     .nock('https://api.heroku.com', (api) => api.delete(`/apps/${COMPUTE_ENV_NAME}`).reply(200))
     .nock('https://api.heroku.com', (api) => api.get(`/apps/${COMPUTE_ENV_NAME}`).reply(200))
     .do(() => {
-      sandbox.stub(Utils, 'resolveOrg' as any).throws('Attempted to resolve an org without a valid org ID');
+      sandbox.stub(Utils, 'findOrgExpirationStatus' as any).throws('This function should not have been called');
     })
-    .add('projectResolveStub', () => {
-      return sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+    .add('findOrgStub', () => {
+      return sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
     })
     .finally(() => {
       sandbox.restore();
@@ -96,7 +117,7 @@ describe('env:delete', () => {
       expect(output).to.include(`Deleting environment ${COMPUTE_ENV_NAME}... done`);
       // If they're deleting the environment after deleting the org, the command shouldn't attempt
       // to clean up function references, and therefore it shouldn't attempt to resolve the project
-      expect(ctx.projectResolveStub).to.not.have.been.called;
+      expect(ctx.findOrgStub).to.not.have.been.called;
     });
 
   test
@@ -104,7 +125,7 @@ describe('env:delete', () => {
     .nock('https://api.heroku.com', (api) => api.get(`/apps/${COMPUTE_ENV_NAME}`).reply(404))
     .do(() => {
       sandbox.stub(Utils, 'resolveOrg' as any).returns({});
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       sandbox.stub(Utils, 'fetchOrg' as any).returns(ORG_MOCK);
     })
     .finally(() => {
@@ -144,7 +165,7 @@ describe('env:delete', () => {
     .stderr()
     .do(() => {
       sandbox.stub(Utils, 'resolveOrg' as any).returns(ORG_MOCK);
-      sandbox.stub(SfdxProject, 'resolve' as any).returns(PROJECT_MOCK);
+      sandbox.stub(SfProject, 'resolve' as any).returns(PROJECT_MOCK);
       sandbox.stub(AliasAccessor.prototype, 'getValue').returns(COMPUTE_ENV_NAME);
       sandbox.stub(Utils, 'findOrgExpirationStatus' as any).returns(false);
     })